edit funktsioon töötab

+++ table.php
@@ -11,6 +11,13 @@
 		deleteCar($_GET["delete"]);
 		
 	}
+	
+	// salvestan muudatuse andmebaasi
+	if(isset($_POST["save"])){
+		
+		updateCar($_POST["id"], $_POST["number_plate"], $_POST["color"]);
+		
+	}
 
 	// käivitan funktsiooni
 	$array_of_cars = getCarData();
@@ -28,6 +35,7 @@
 		<th>värv</th>
 		<th>user_id</th>
 		<th>kustuta</th>
+		<th>edit</th>
 	</tr>
 	
 	<?php
@@ -35,13 +43,37 @@
 		// massiivi pikkus count()
 		for($i = 0;$i < count($array_of_cars);$i++){
 			//echo $array_of_cars[$i]->id;
-			echo "<tr>";
-			echo "<td>".$array_of_cars[$i]->id."</td>";
-			echo "<td>".$array_of_cars[$i]->plate."</td>";
-			echo "<td>".$array_of_cars[$i]->color."</td>";
-			echo "<td>".$array_of_cars[$i]->user_id."</td>";
-			echo "<td><a href='?delete=".$array_of_cars[$i]->id."'>X</a></td>";
-			echo "</tr>";
+			
+			// kasutaja tahab muuta seda rida
+			if(isset($_GET["edit"]) && $array_of_cars[$i]->id == $_GET["edit"]){
+				
+				echo "<tr>";
+				echo "<form action='table.php' method='post'>";
+				echo "<input type='hidden' name='id' value='".$array_of_cars[$i]->id."'>";
+				echo "<td>".$array_of_cars[$i]->id."</td>";
+				echo "<td><input name='number_plate' value='".$array_of_cars[$i]->plate."'></td>";
+				echo "<td><input name='color' value='".$array_of_cars[$i]->color."'></td>";
+				echo "<td>".$array_of_cars[$i]->user_id."</td>";
+				echo "<td><a href='table.php'>cancel</a></td>";
+				echo "<td><input type='submit' name='save'></td>";
+				echo "</form>";
+				echo "</tr>";
+				
+				
+			}else{
+				echo "<tr>";
+				echo "<td>".$array_of_cars[$i]->id."</td>";
+				echo "<td>".$array_of_cars[$i]->plate."</td>";
+				echo "<td>".$array_of_cars[$i]->color."</td>";
+				echo "<td>".$array_of_cars[$i]->user_id."</td>";
+				echo "<td><a href='?delete=".$array_of_cars[$i]->id."'>X</a></td>";
+				echo "<td><a href='?edit=".$array_of_cars[$i]->id."'>edit</a></td>";
+				echo "</tr>";
+				
+				
+			}
+			
+		
 		}
